+++ Bin/Status.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Hoa
  *
@@ -44,9 +46,6 @@ use Hoa\Worker;
  * Class \Hoa\Worker\Bin\Status.
  *
  * Status all workers.
- *
- * @copyright  Copyright © 2007-2017 Hoa community
- * @license    New BSD License
  */
 class Status extends Console\Dispatcher\Kit
 {
@@ -65,8 +64,6 @@ class Status extends Console\Dispatcher\Kit
 
     /**
      * The entry method.
-     *
-     * @return  void
      */
     public function main()
     {
@@ -203,8 +200,6 @@ class Status extends Console\Dispatcher\Kit
 
     /**
      * The command usage.
-     *
-     * @return  void
      */
     public function usage()
     {
